third_party/blink/web_tests/fast/canvas/check-stale-putImageData-expected.txt

CONSOLE MESSAGE: a == b? false
CONSOLE MESSAGE: a == c? false
